implementei internacionalização

+++ src/main/java/sgcp/service/ContratoService.java
@@ -1,8 +1,11 @@
 package sgcp.service;
 
 import java.util.List;
+import java.util.Locale;
 
+import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import sgcp.model.Contrato;
 import sgcp.model.ContratoKey;
@@ -12,39 +15,64 @@ import sgcp.repository.ContratoRepository;
 public class ContratoService {
 
 	private final ContratoRepository cr;
+
+	private final MessageSource messages;
 	
-	public ContratoService(ContratoRepository cr) {
+	public ContratoService(ContratoRepository cr, MessageSource messages) {
 		
 		this.cr = cr;
+		this.messages = messages;
 	}
 	
 	public List<Contrato> listaContratos(){
 		return cr.findAll();
 	}
 
-	public Contrato incluirContrato(Contrato contrato) {
+	public String incluirContrato(Contrato contrato, Locale locale) {
 		
+		String responseMessage = null;
 		
-		System.out.println("Contrato incluído com sucesso");
-		return cr.save(contrato);
+		if(!StringUtils.isEmpty(contrato)) {
+
+			Contrato c = cr.save(contrato);			
+			responseMessage = String.format(messages.getMessage("contrato.create.message",null,locale), c.toString());
+		}
+
+		System.out.println("Contrato incluído com sucesso");	
+		return responseMessage;		
+	}
+	
+	public String alterarContrato(Contrato contrato, Locale locale) {
 		
+		String responseMessage = null;
+		
+		if(!StringUtils.isEmpty(contrato)) {
+
+			Contrato c = cr.save(contrato);			
+			responseMessage = String.format(messages.getMessage("contrato.update.message",null,locale), c.toString());
+		}
+
+		System.out.println("Contrato alterado com sucesso");	
+		return responseMessage;		
 	}
 
-	public void exluirContrato(ContratoKey ck) {
+	public String exluirContrato(ContratoKey ck, Locale locale) {
 		
-		cr.deleteById(ck);
+		String responseMessage = null;
 		
-	}
+		if(!StringUtils.isEmpty(ck)) {
 
-	public Contrato alterarContrato(Contrato contrato) {
-		
-		return cr.save(contrato);
+			cr.deleteById(ck);			
+			responseMessage = String.format(messages.getMessage("contrato.delete.message",null,locale), ck.toString());
+		}
+
+		System.out.println("Contrato excluído com sucesso");	
+		return responseMessage;				
 	}
 
 	public Contrato detalharContrato(ContratoKey ck) {
 		
-		Contrato contrato =  cr.getById(ck);
-		
+		Contrato contrato =  cr.getById(ck);		
 		return contrato;
 	}
 }
